Make kazoe yakuman configurable
Create a new boolean kazoeyakuman that enables kazoe yakuman rules. It
defaults to true.

Kazoe yakuman ari enables the player to score a yakuman with a hand
worth 13 han or more. When nashi, hands worth 13+ han are only worth
sanbaiman.

This toggle adjusts the appearance of the limit score headings. If
false, sanbaiman is listed as 11+ and yakuman unatainable by han. If
true, sanbaiman is listed as 11-12 and yakuman as 13+.